Update README.asciidoc
Gentoo has a wiki page documenting Kakoune installation. The method given here ("emerge kakoune") will not work on a default Gentoo installation, as Kakoune is currently in the Testing branch rather than in the Stable branch. A link to the documentation explaining how to install Kakoune on Gentoo seems pertinent.