Add black-on-white colorscheme
Kakoune doesn't have an equivalent of Vim's ":syntax off". This
colorscheme is as close as it gets: it uses various shades of grey
for cursors, selections and completion/info boxes, and black/white for
everything else.  The high-contrast makes it ideal for bright settings.

I also have the dual white-on-black colorscheme but it's not tested
much, so I left it on my "colorschemes" branch.